Modified plugin to test the impacts of the changes brought in the Christian's plugin with proportional min & max scenarios.
the thresholds of the magnitude of the imbalance is set to 1 intead of 100.

The method is set to proportional in the code.
The threshold method is set to share in the code

The new options are:
New stocks position (TRUE/FALSE):TRUE=move the stocks adjustment after the proportional balancing

Thresholds (TRUE/FALSE): TRUE=use the relaxed thresholds (0% for feed, 10% for food, feed and losses and 100% for industrial uses)

new food residual (TRUE/FALSE): TRUE=use the modified approche of food residual (food the only utilization and the commodity has historique value for food)